requirements.txt
streamlit>=1.32.0
joblib
numpy
pandas
scikit-learn
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.1.0
langgraph>=0.1.0
huggingface-hub>=0.22.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
fpdf2>=2.7.0
pypdf>=3.0.0
